requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.8.0
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
